refactor(M20DS-2): move palette config, restore grey tokens and expose theme colors
