+++ src/theme/foundations/colors.ts
@@ -10,14 +10,6 @@ const neutral = {
     100: '#E8EBF1',
     50: '#F4F5F8',
   },
-  850: '#2B2E38',
-  700: '#555C70',
-  650: '#636B82',
-  450: '#99A3C1',
-  300: '#BBC2D6',
-  200: '#D2D6E3',
-  100: '#E8EBF1',
-  50: '#F4F5F8',
   white: '#FFFFFF',
 };
